backend/pyproject.toml
[tool.poetry]
name = "seculite-backend"
version = "0.1.0"
description = "Backend for SecuLite v2"
authors = ["Patrick Böhme <pa.boe90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
django = ">=5.0,<5.1"
djangorestframework = ">=3.16.0,<4.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
celery = ">=5.5.2,<6.0.0"
redis = ">=6.0.0,<7.0.0"
gunicorn = ">=23.0.0,<24.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
dj-database-url = ">=2.3.0,<3.0.0"
django-cors-headers = ">=4.7.0,<5.0.0"
djangorestframework-simplejwt = ">=5.5.0,<6.0.0"
django-filter = ">=25.1,<26.0"
drf-yasg = ">=1.21.10,<2.0.0"
dj-rest-auth = ">=6.0.0,<7.0.0"
django-allauth = ">=0.63.0,<0.64.0"
requests = ">=2.30.0,<3.0.0"
django-celery-beat = ">=2.6.0,<3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
